+++ fronts-client/src/components/Editions/ManageEdition.tsx
@@ -12,7 +12,7 @@ import {
   createIssue
 } from 'services/editionsApi';
 import { withRouter, RouteComponentProps } from 'react-router';
-import Spinner from 'shared/components/async/Spinner';
+import Spinner from 'components/async/Spinner';
 import InformationMsg from 'components/alert/InformationMsg';
 import url from 'constants/url';
