+++ fronts-client/src/components/article/DraggableArticleImageContainer.tsx
